Move bubble elements to their own file
